requirements.txt
requests
beautifulsoup4
chromadb
unstructured
langchain
langchain-community
langchain-chroma
azure-ai-inference
python-dotenv
pytest